CLEANUP: Migrated `privacyByDefault.spec.ts` into `privacy.spec.ts`
